@@ -22,7 +22,7 @@ import org.springframework.security.authentication.AuthenticationTrustResolver
 import org.springframework.security.core.Authentication
 import org.springframework.security.core.context.SecurityContextHolder as SCH
 
-import javax.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletRequest
 
 /**
  * Utility methods.
