@@ -14,9 +14,9 @@
  */
 package grails.plugin.springsecurity.web.access
 
-import javax.servlet.ServletException
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.servlet.ServletException
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 
 import org.springframework.beans.factory.InitializingBean
 import org.springframework.security.access.AccessDeniedException
