@@ -14,8 +14,8 @@
  */
 package grails.plugin.springsecurity.web.authentication
 
-import javax.servlet.http.HttpServletRequest
-import javax.servlet.http.HttpServletResponse
+import jakarta.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletResponse
 
 import org.springframework.security.core.Authentication
 import org.springframework.security.web.authentication.NullRememberMeServices
