@@ -3,7 +3,7 @@ package grails.plugin.springsecurity.web.filter
 import groovy.transform.CompileStatic
 import org.springframework.util.Assert
 
-import javax.servlet.http.HttpServletRequest
+import jakarta.servlet.http.HttpServletRequest
 
 @CompileStatic
 class HttpMethodOverrideDetector {
